+++ yash-syntax/src/parser/lex/dollar.rs
@@ -144,7 +144,7 @@ mod tests {
             assert_eq!(location.code.start_line_number.get(), 1);
             assert_eq!(location.code.source, Source::Unknown);
             assert_eq!(location.range, 0..3);
-            assert_eq!(content, "");
+            assert_eq!(&*content, "");
         });
         assert_eq!(block_on(lexer.peek_char()), Ok(None));
 
@@ -159,7 +159,7 @@ mod tests {
             assert_eq!(location.code.start_line_number.get(), 1);
             assert_eq!(location.code.source, Source::Unknown);
             assert_eq!(location.range, 0..12);
-            assert_eq!(content, " foo bar ");
+            assert_eq!(&*content, " foo bar ");
         });
         assert_eq!(block_on(lexer.peek_char()), Ok(None));
     }